# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "INVENTRAK Inventory Management API",
    "description": "Backend API for the INVENTRAK inventory management system. Both the SQLite (Express) backend and the npm-free fallback expose this identical contract. Live docs: GET /api/docs",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4001",
      "description": "Local backend"
    }
  ],
...
  "paths": {
...
    "/api/sales": {
      "get": {
        "tags": [
          "Sales"
        ],
        "operationId": "listSales",
        "summary": "List sales transactions (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sales (plain array, or { data, pagination } when page/limit provided)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Sale"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/PaginatedSales"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Sales"
        ],
        "operationId": "createSale",
        "summary": "Record a sale",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaleInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sale recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok",
                    "total"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "total": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Product not found or inactive",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation failed or malformed input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Admin access required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ],
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
...
      "PaginatedSales": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Sale"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
      "SaleInput": {
        "type": "object",
        "required": [
          "product_id",
          "qty"
        ],
        "properties": {
          "product_id": {
            "type": "integer",
            "minimum": 1
          },
          "qty": {
            "type": "number",
            "minimum": 0.01
          },
          "customer_name": {
            "type": "string"
          }
        }
      },
      "Sale": {
        "type": "object",
        "required": [
          "id",
          "product_id",
          "qty",
          "unit_price",
          "total_amount",
          "transaction_date"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "product_id": {
            "type": "integer"
          },
          "qty": {
            "type": "number"
          },
          "unit_price": {
            "type": "number"
          },
          "total_amount": {
            "type": "number"
          },
          "transaction_date": {
            "type": "string"
          },
          "customer_name": {
            "type": "string"
          },
          "customer_id": {
            "type": "integer",
            "nullable": true,
            "description": "Links to the Customer Record this sale belongs to. Null on seeded/historical rows, which carry only a free-text name."
          },
          "product_name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}